refactor: :art: module system refactor
