+++ src/lib.rs
@@ -1,6 +1,3 @@
 pub mod elliptic_curve;
-pub mod elliptic_curve_point;
-pub mod finite_field;
-pub mod finite_field_element;
 pub mod secp256k1;
 pub mod utils;
